make pixel coordinates integers change Agent.move to use getNextPosition

+++ Pacman/src/models/Game.java
@@ -96,22 +96,22 @@ public class Game {
 	}
 
 	public boolean checkMove(Agent agent, DirectionEnum direction) {
-		double[] nextPosition = agent.getNextPosition(direction);
+		int[] nextPosition = agent.getNextPosition(direction);
 		return (isPassable(nextPosition[0], nextPosition[1]));
 	}
 
 	//check if pac man can go to that place
 	//is it as simple as checking if position is a wall, or do we need to enforce min distance from a wall?
 	//in tunnel of 1 square width, pacman shouldn't be able to move toward either wall
-	public boolean isPassable(double x, double y){
+	public boolean isPassable(int x, int y){
 		int[] intPosition = doubleToIntPosition(x, y);
 		return maze[intPosition[0]][intPosition[1]] != TileEnum.Wall;
 	}
 
 	//transate the double position to closest integer position in the maze
 	//is this the right way up? not sure which way x and y go in the array
-	public int[] doubleToIntPosition(double x, double y){
-		int[] intPoint = {(int) x/16, (int) y/16};
+	public int[] doubleToIntPosition(int x, int y){
+		int[] intPoint = { x/16, y/16};
 		return intPoint;
 	}
